Fix null bitmap building to allow for more than 8 columns
Previously, the code as it was would only write a single byte to
the packet buffer. Which can only represent 8 bits, and thus
columns. So far as I can tell, MySQL can support up to 4096 columns
which would require 512 bytes for the bitmap. This change allows
for that by filling out the bitmap and writing it to the packet
buffer a byte at a time.

A better way to handle this would probably be to ask the buffer code
to "reserve" bytes for us in a single allocation. From which we
could just fill in the bits along that pre-allocated memory in one
shot. But that will require some more thought for the buffer API.
So I'll save it for another day :)